requirements.txt
click>=8.0.0
notional>=0.3.0
dateparser>=1.1.0
python-dotenv>=1.0.0
requests>=2.25.0